Login fixed Register

+++ Frontend/src/app/components/register/register.component.ts
@@ -6,6 +6,7 @@ import { Router } from '@angular/router';
 import { HttpErrorResponse } from '@angular/common/http';
 import { JwtService } from 'src/app/services/jwt.service';
 import { User } from 'src/app/interfaces/user';
+import { HotToastService } from '@ngneat/hot-toast';
 @Component({
   selector: 'app-register',
   templateUrl: './register.component.html',
@@ -26,7 +27,8 @@ export class RegisterComponent implements OnInit {
   constructor(
     private auth: AuthusersService,
     private router: Router,
-    private jwt: JwtService
+    private jwt: JwtService,
+    private toast : HotToastService
   ) {}
 
   ngOnInit(): void {}
@@ -49,32 +51,50 @@ export class RegisterComponent implements OnInit {
   step = 1;
 
   nextStep(form: FormGroup) {
-    this.step = 2;
+    if(this.step == 1){
+      this.step = 2;
+    }else
+    {
+      this.step = 1;
+    }
+
+
   }
 
   onRegister(form: FormGroup) {
-    this.auth.postData(form.value).subscribe(
-      (results: any) => {
-        this.auth.saveToken(results.token);
-        this.user = this.jwt.getData(results.token);
 
-        if (this.user != null) {
-          sessionStorage.setItem('role', this.user.account);
-          this.role = this.user.account;
+    this.toast.loading('Signing you up ...',{duration:200})
+
+    if(form.value.password === form.value.confirmPassword){
+      this.auth.postData(form.value).subscribe(
+        (results: any) => {
+          this.auth.saveToken(results.token);
+          this.user = this.jwt.getData(results.token);
+
+          if (this.user != null) {
+            sessionStorage.setItem('role', this.user.account);
+            this.role = this.user.account;
+            this.toast.success('Welcome to Skoolify')
+          }
+
+          if (this.role == 'PARENT') {
+            this.router.navigateByUrl('/parent-home');
+          } else if (this.role == 'OWNER') {
+            this.router.navigateByUrl('/owner-home');
+          } else if (this.role == 'ADMIN') {
+            this.router.navigateByUrl('/admin/schools');
+          }
+        },
+
+        (error: HttpErrorResponse) => {
+          this.toast.error(error.error.message)
+
         }
+      );
 
-        if (this.role == 'PARENT') {
-          this.router.navigateByUrl('/parent-home');
-        } else if (this.role == 'OWNER') {
-          this.router.navigateByUrl('/owner-home');
-        } else if (this.role == 'ADMIN') {
-          this.router.navigateByUrl('/admin/schools');
-        }
-      },
+    }else{
+      this.toast.warning('Oops! Passwords do not match');
+    }
 
-      (error: HttpErrorResponse) => {
-        console.log(error);
-      }
-    );
   }
 }
